Only delete IA if it has an instant_article_id

+++ bulbs/content/models.py
@@ -621,46 +621,47 @@ def content_deleted(sender, instance=None, **kwargs):
 
 def delete_from_instant_article_api(sender, instance=None, **kwargs):
     if getattr(settings, 'FACEBOOK_POST_TO_IA', False):
-        fb_api_url = getattr(settings, 'FACEBOOK_API_BASE_URL', None)
-        fb_token_path = getattr(settings, 'FACEBOOK_TOKEN_VAULT_PATH', None)
+        if getattr(instance, 'instant_article_id', None):
+            fb_api_url = getattr(settings, 'FACEBOOK_API_BASE_URL', None)
+            fb_token_path = getattr(settings, 'FACEBOOK_TOKEN_VAULT_PATH', None)
 
-        if not fb_api_url or not fb_token_path:
-            logger.error('''
-                Error in Django Settings.\n
-                FACEBOOK_API_BASE_URL: {0}\n
-                FACEBOOK_TOKEN_VAULT_PATH: {1}'''.format(fb_api_url,
-                                                         fb_token_path))
-            return
+            if not fb_api_url or not fb_token_path:
+                logger.error('''
+                    Error in Django Settings.\n
+                    FACEBOOK_API_BASE_URL: {0}\n
+                    FACEBOOK_TOKEN_VAULT_PATH: {1}'''.format(fb_api_url,
+                                                             fb_token_path))
+                return
 
-        fb_access_token = vault.read(fb_token_path).get('authtoken')
-        if fb_access_token is None:
-            logger.error('Missing FB Auth Token in Vault.\n')
-            return
+            fb_access_token = vault.read(fb_token_path).get('authtoken')
+            if fb_access_token is None:
+                logger.error('Missing FB Auth Token in Vault.\n')
+                return
 
-        delete = requests.delete('{0}/{1}?access_token={2}'.format(
-            fb_api_url,
-            instance.instant_article_id,
-            fb_access_token
-        ))
+            delete = requests.delete('{0}/{1}?access_token={2}'.format(
+                fb_api_url,
+                instance.instant_article_id,
+                fb_access_token
+            ))
 
-        if not delete.ok:
-            logger.error('''
-                Error in deleting Instant Article.\n
-                Content ID: {0}\n
-                IA ID: {1}\n
-                Status Code: {2}
-                Request: {3}'''.format(instance.id,
-                                       instance.instant_article_id,
-                                       delete.status_code,
-                                       delete.__dict__))
-        else:
-            status = delete.json().get('success')
-            if bool(status) is not True:
+            if not delete.ok:
                 logger.error('''
                     Error in deleting Instant Article.\n
                     Content ID: {0}\n
                     IA ID: {1}\n
-                    Error: {2}'''.format(instance.id, instance.instant_article_id, delete.json()))
+                    Status Code: {2}
+                    Request: {3}'''.format(instance.id,
+                                           instance.instant_article_id,
+                                           delete.status_code,
+                                           delete.__dict__))
+            else:
+                status = delete.json().get('success')
+                if bool(status) is not True:
+                    logger.error('''
+                        Error in deleting Instant Article.\n
+                        Content ID: {0}\n
+                        IA ID: {1}\n
+                        Error: {2}'''.format(instance.id, instance.instant_article_id, delete.json()))
 
 ##
 # signal hooks
